EHN move backend test in testing modules
